requirements.txt
tensorflow==2.10.1
gdown==4.6.0
pillow==9.4.0
numpy==1.22.4
streamlit
flask